+++ src/main/java/com/ccc/api/repository/MetricAlarmRepository.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 import com.ccc.api.model.MetricAlarm;
 
 @Repository
-public interface MetricAlarmRepository extends JpaRepository<MetricAlarm, Integer> {
+public interface MetricAlarmRepository extends JpaRepository<MetricAlarm, Long> {
 }
